Restored to version 1.2.0

+++ CLUEstering/binding.cc
@@ -1,4 +1,3 @@
-#include <cstdint>
 #include <string>
 #include <vector>
 #include <iostream>
@@ -23,13 +22,13 @@ struct Points {
 	std::vector<float> rho;
 	std::vector<float> delta;
 	std::vector<int> nearestHigher;
-	std::vector<uint16_t> clusterIndex;
+	std::vector<int> clusterIndex;
 	std::vector<std::vector<int>> followers;
-	std::vector<uint8_t> isSeed; // the elements should only be 0 or 1
+	std::vector<int> isSeed;
 	int n;
 
 	void clear() {
-		for(uint8_t i = 0; i != Ndim; ++i) {
+		for(int i = 0; i != Ndim; ++i) {
 			coordinates_[i].clear();
 		}
 		weight.clear();
@@ -71,7 +70,7 @@ public:
 	int getGlobalBin(std::vector<float> coords) const {
 		int globalBin = getBin(coords[0],0);
 		int nTilesPerDim = std::pow(nTiles,1.0/Ndim);
-      for(uint8_t i = 1; i != Ndim; ++i) {
+      for(int i = 1; i != Ndim; ++i) {
 			globalBin += nTilesPerDim*getBin(coords[i],i);
       }
       return globalBin;
@@ -80,7 +79,7 @@ public:
 	int getGlobalBinByBin(std::vector<int> Bins) const {
 		int globalBin = Bins[0];
       int nTilesPerDim = std::pow(nTiles,1.0/Ndim);
-      for(uint8_t i = 1; i != Ndim; ++i) {
+      for(int i = 1; i != Ndim; ++i) {
 			globalBin += nTilesPerDim*Bins[i];
       }
       return globalBin;
@@ -94,7 +93,7 @@ public:
 		auto cellsSize = coordinates[0].size();
       for(int i = 0; i < cellsSize; ++i) {
 			std::vector<float> bin_coords;
-			for(uint8_t j = 0; j != Ndim; ++j) {
+			for(int j = 0; j != Ndim; ++j) {
 				bin_coords.push_back(coordinates[j][i]);
 			} 
 			tiles_[getGlobalBin(bin_coords)].push_back(i);
@@ -104,7 +103,7 @@ public:
 	std::array<int,2*Ndim> searchBox(std::array<std::vector<float>,Ndim> minMax_){   // {{minX,maxX},{minY,maxY},{minZ,maxZ},...}
 		std::array<int, 2*Ndim> minMaxBins;
       int j = 0;
-      for(uint8_t i = 0; i != Ndim; ++i) {
+      for(int i = 0; i != Ndim; ++i) {
 			minMaxBins[j] = getBin(minMax_[i][0],i);
 			minMaxBins[j+1] = getBin(minMax_[i][1],i);
 			j += 2;
@@ -131,7 +130,7 @@ public:
 template <uint8_t Ndim>
 class ClusteringAlgo{
 public:
-	ClusteringAlgo(float dc, float rhoc, float outlierDeltaFactor, uint8_t pPBin) {
+	ClusteringAlgo(float dc, float rhoc, float outlierDeltaFactor, int pPBin) {
 		dc_ = dc; 
 		rhoc_ = rhoc;
 		outlierDeltaFactor_ = outlierDeltaFactor;
@@ -143,7 +142,7 @@ public:
 	float dc_;  // cut-off distance in the calculation of local density
 	float rhoc_;  // minimum density to promote a point as a seed or the maximum density to demote a point as an outlier
 	float outlierDeltaFactor_;
-	uint8_t pointsPerTile_; // average number of points found in a tile
+	int pointsPerTile_; // average number of points found in a tile
     
 	Points<Ndim> points_;
   
@@ -151,7 +150,7 @@ public:
 		//points_.clear();
 		// input variables
 		for(int i = 0; i < n; ++i) {
-			for(uint8_t j = 0; j != Ndim; ++j) {
+			for(int j = 0; j != Ndim; ++j) {
 				points_.coordinates_.push_back({});
 				points_.coordinates_[j].push_back(coordinates[j][i]);
 			}
@@ -191,7 +190,7 @@ public:
 		std::array<float,Ndim> tileSizes;
 		int NperDim = std::pow(NTiles,1.0/Ndim);
 
-		for(uint8_t i = 0; i != Ndim; ++i) {
+		for(int i = 0; i != Ndim; ++i) {
 			float tileSize;
 			float dimMax = *std::max_element(points_.coordinates_[i].begin(),points_.coordinates_[i].end());
 			float dimMin = *std::min_element(points_.coordinates_[i].begin(),points_.coordinates_[i].end());
@@ -303,7 +302,7 @@ private:
 	for (int i = 0; i < points_.n; ++i){
       // push index of points into tiles
       std::vector<float> coords;
-      for(uint8_t j = 0; j != Ndim; ++j) {
+      for(int j = 0; j != Ndim; ++j) {
         coords.push_back(points_.coordinates_[j][i]);
       }
       tiles.fill(coords, i);
@@ -315,7 +314,7 @@ private:
     for(int i = 0; i < points_.n; ++i) {
       // get search box
       std::array<std::vector<float>,Ndim> minMax;
-      for(uint8_t j = 0; j != Ndim; ++j) {
+      for(int j = 0; j != Ndim; ++j) {
         std::vector<float> partial_minMax{points_.coordinates_[j][i]-dc_,points_.coordinates_[j][i]+dc_};
         minMax[j] = partial_minMax;
       }
@@ -350,7 +349,7 @@ private:
 
       // get search box
       std::array<std::vector<float>,Ndim> minMax;
-      for(uint8_t j = 0; j != Ndim; ++j) {
+      for(int j = 0; j != Ndim; ++j) {
         std::vector<float> partial_minMax{points_.coordinates_[j][i]-dm,points_.coordinates_[j][i]+dm};
         minMax[j] = partial_minMax;
       }
@@ -434,7 +433,7 @@ private:
 
 	inline float distance(int i, int j) const {
 		float qSum = 0;   // quadratic sum
-		for(uint8_t k = 0; k != Ndim; ++k) {
+		for(int k = 0; k != Ndim; ++k) {
 			qSum += std::pow(points_.coordinates_[k][i] - points_.coordinates_[k][j],2);
 		}
     
@@ -446,7 +445,7 @@ private:
 /////////////////////
 //////  Run.h  //////
 /////////////////////
-std::vector<std::vector<int>> run1(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run1(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<1> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -454,7 +453,7 @@ std::vector<std::vector<int>> run1(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run2(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run2(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<2> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -462,7 +461,7 @@ std::vector<std::vector<int>> run2(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 			
-std::vector<std::vector<int>> run3(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run3(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<3> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -470,7 +469,7 @@ std::vector<std::vector<int>> run3(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run4(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run4(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<4> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -478,7 +477,7 @@ std::vector<std::vector<int>> run4(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run5(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run5(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates,	std::vector<float> const& weight) {
 	ClusteringAlgo<5> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -486,7 +485,7 @@ std::vector<std::vector<int>> run5(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run6(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run6(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<6> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -494,7 +493,7 @@ std::vector<std::vector<int>> run6(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run7(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run7(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<7> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -502,7 +501,7 @@ std::vector<std::vector<int>> run7(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run8(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run8(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<8> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -510,7 +509,7 @@ std::vector<std::vector<int>> run8(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run9(float dc, float rhoc, float outlier, uint8_t pPBin,  
+std::vector<std::vector<int>> run9(float dc, float rhoc, float outlier, int pPBin,  
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<9> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -518,7 +517,7 @@ std::vector<std::vector<int>> run9(float dc, float rhoc, float outlier, uint8_t 
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> run10(float dc, float rhoc, float outlier, uint8_t pPBin, 
+std::vector<std::vector<int>> run10(float dc, float rhoc, float outlier, int pPBin, 
 		std::vector<std::vector<float>> const& coordinates, std::vector<float> const& weight) {
 	ClusteringAlgo<10> algo(dc,rhoc,outlier,pPBin);
 	algo.setPoints(coordinates[0].size(), coordinates, weight);
@@ -526,8 +525,8 @@ std::vector<std::vector<int>> run10(float dc, float rhoc, float outlier, uint8_t
 	return algo.makeClusters();
 }
 
-std::vector<std::vector<int>> mainRun(float dc, float rhoc, float outlier, uint8_t pPBin, 
-            std::vector<std::vector<float>> const& coords, std::vector<float> const& weight, uint8_t Ndim) {
+std::vector<std::vector<int>> mainRun(float dc, float rhoc, float outlier, int pPBin, 
+            std::vector<std::vector<float>> const& coords, std::vector<float> const& weight, int Ndim) {
     // Running the clustering algorithm //
 	if (Ndim == 1) {
 		return run1(dc,rhoc,outlier,pPBin,coords,weight);
